Split function I believe to be general into a separate file utils.py
